Significant rework of get and post handlers. This will break existing sites, but it's worth it.
This change set causes the get and post handlers to create functions,
which gives them access to their lexical context. Previously, handlers
were evaluated in the context of (NunjaRequestHandler handleRequest:),
which simplified a few things but made it difficult to declare helper
objects, functions, and values for use in action handlers.

Now get and post implicitly declare functions that are called with
three arguments: MATCH, REQUEST, and RESPONSE. MATCH is the result
of a regex match on the action path and was previously available
through the @match instance variable.  To set the TITLE and HEAD
of the response, set the corresponding key-values in the RESPONSE
dictionary.

See the sample/site.nu for examples.

This revision contains enough of the old code that it should still
to use the new Nunja with old sites if you set the $do-it-the-old-way
global to something non-nil, but this will disappear soon.