Preliminary support for asynchronous dns and http queries.
These additions are to allow Nunja-based web applications
to call out to other servers asynchronously. Requests
made using the internal API are accompanied by blocks
that specify actions to be performed when queries
complete.

All new interfaces are preliminary and are likely
to change. See sample/site.nu for usage.